requirements.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication
PyJWT==2.8.0
djangorestframework-simplejwt==5.3.1

# Caching & Async
redis==5.0.1
django-redis==5.4.0
celery==5.3.6

# Vector Database
qdrant-client==1.7.3

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
python-magic==0.4.27

# Embeddings & LLM
openai==1.10.0
sentence-transformers==2.3.1
tiktoken==0.5.2

# MFA & QR Codes
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.2.0

# Utilities
python-dotenv==1.0.0
python-decouple==3.8
Pillow==10.2.0
sentry-sdk==2.23.1

# Rate Limiting
django-ratelimit==4.1.0

# Validation
pydantic==2.5.3

# Development
django-extensions==3.2.3
ipython==8.20.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Testing
pytest==7.4.4
pytest-django==4.7.0
factory-boy==3.3.0
faker==22.0.0
